clean ups, standrad error code, action structure matches reducer structure
