+++ Mobile/src/actions/index.js
@@ -1,5 +1,6 @@
 export * from './actions_auth';
 export * from './actions_features';
 export * from './actions_feedback';
+export * from './actions_projects';
 export * from './actions_solutions';
 export * from './actions_user';
